Fixed PIC Slave bug
This fixes an issue with the creation of an interrupt mask
for the slave PIC. The IRQ value being used to generate the mask
was not relative to the slave, and so resulted in shifting values
too far, and generating clear masks (all 1's). This disabled all
interrupts for the slave PIC instead of enabling certain ones.