Makefile: generify tools so they're either standard or overridable
Some users will want to specify a compiler/linker other than the default
gcc/ld, e.g. if you want to use a cross-compiler or a specific version
of the aforementioned tools. Use standard CC variable for compiler and
LD for linker.

`truncate` isn't available on all platforms, so use `head` to truncate
the image to floppy disk size.

Some platforms don't use genisoimage, so also make that overridable.

For now, it's important that the subsituted tool takes the same
arguments, but it's definitely possible that we'll find tools in the
later phases (truncation and iso generation) that take different
arguments or have small pain points.